Extend loan protocol with dynamic liquidation
Give the lender the ability to liquidate the collateral if they can
prove that its value has fallen below a predetermined threshold. The
proof is composed of a message and a signature on the message from a
known, trusted oracle.

The oracle API is only loosely defined here, in the form of a
crate-local module used for testing.