Add COV_SCRIPT_BYTES constant
AFAIK, we cannot construct a `const` Script, so we must use bytes.

It is a bit annoying to have to construct the script just to pass it
into `descriptor_satisfiers` as an argument, but otherwise we end up
returning a value referencing the local variable for the covenant
script inside `descriptor_satisfiers`.

Alternatives are welcome.