Deprecate more restrictive API on Lender0
Users of the library want to be able to control how and what they
serialize. Before this change they had to live with passing around the
`LoanRequest` object, but now they are free to decompose it and pass
the individual fields to the new `build_loan_transaction` API on
`Lender0`.

Co-authored-by: Daniel Karzel <daniel@comit.network>