Add units indicator to Oracle message
For price-signalled liquidation to work properly, the consumer
of the oracle message(s) must know what exchange-rate units the
oracle is using. This PR adds a filed "units" the JSON response
and in the documentation (comments in oracle.rs) indicates that
the sample oracle is using millicents. Values remain integers.

Additionally, a typo was discovered in Rocket.toml that has been
fixed up.

Resolves: #53