Document test wallet money-printing behaviour
The test wallet creates UTXOs out of thin air during coin selection.
We use these UTXOs as inputs in our protocol transactions.

When verifying these transactions, the UTXOs that we created are
assumed to be correct: that is, their previous outputs are not
verified (which is why we're creating coins out of thin air).

Therefore, when creating these UTXOs the value of the nonexistent
previous outputs can be anything. We set it to zero to stress the fact
that money is being created.

Conversely, the explicit asset ID must be part of the list of
nonexistent previous outputs in order to produce a valid surjection
proof.