Ensure that transaction assets, amounts and proofs are valid
When moving away from blockchain-node based tests we unknowingly relaxed
the requirements on our transactions, which was a problem.
A transaction may correctly spend its inputs and still be invalid by
virtue of being inflationary (or deflationary).

We have now added checks to ensure that all transactions have correct
assets, amounts and proofs.