Do not attempt to construct 0-amount TxOuts
Change outputs are now only added to the loan transaction if
necessary. This fixes bug
https://github.com/comit-network/baru/issues/38.

We therefore no longer need to always add a buffer to the amount we
select coins for. The original loan protocol tests now use the least
number of outputs possible (we still need to a collateral change
output for the loan transaction because we don't know the transaction
fee ahead of time. See:
https://github.com/comit-network/baru/discussions/57).

To verify that we still allow the presence of change
outputs (principal change outputs in particular), we've added the test
`can_run_protocol_with_principal_change_outputs`.