Create serializable encrypted wallet type
Created a new EncryptedWallet type specifically for serialisation and
secure storage. The Wallet and EncryptedWallet can be converted to each
other with a encyption_key. The cached Utxo's are lost in this process.
Serde has been removed from the Wallet type as it is encourages leakage
of the cleartext root xprv.

We removed the secret key field from the wallet and derive it everytime
from the root xprv.

Construct wallet using a seed.

Add test that fails if encrypted wallet json format is changed

Error when deprecated API's are used with new ones. This has the added
bonus of allowing as to derive PartialEq on the
Wallet type and assert on it in tests